Fix compatibility issues for ReaderIterator
